use project or global map scales
